% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_convo.R
\name{get_convo}
\alias{get_convo}
\title{Get conversation tweets}
\usage{
get_convo(token, tweet_id = NULL)
}
\arguments{
\item{token}{List. Twitter API tokens.}

\item{tweet_id}{Character string. Tweet id of any tweet that is part of the threaded conversation.}
}
\value{
A dataframe of tweets.
}
\description{
Get conversation tweets
}
\examples{
\dontrun{
# get twitter conversation by tweet id
tweets <- get_convo(token, "1309703812123226112")
}

}
