% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_convos.R
\name{get_convos}
\alias{get_convos}
\title{Get conversation tweets}
\usage{
get_convos(token, tweet_ids = NULL)
}
\arguments{
\item{token}{List. Twitter API tokens.}

\item{tweet_ids}{List. Tweet ids of any tweet that are part of the threaded conversations of interest.}
}
\value{
A dataframe of tweets.
}
\description{
Get conversation tweets
}
\examples{
\dontrun{
# get twitter conversations by tweet ids
urls <- c("https://twitter.com/xxxx/status/xxxxxxxxxxxx",
tweets <- get_convos(token, urls)
}

}
