% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tcs.R
\name{get_tcs}
\alias{get_tcs}
\title{Get conversation tweets}
\usage{
get_tcs(tweet_ids = NULL, token)
}
\arguments{
\item{tweet_ids}{List. Tweet ids of any tweet that are part of the thread conversations of interest. Also accepts
a list of tweet URLs or a mixed list.}

\item{token}{List. Twitter API tokens.}
}
\value{
A dataframe of tweets.
}
\description{
Get conversation tweets
}
\examples{
\dontrun{
# get twitter conversations by tweet ids or urls
tweet_ids <- c("xxxxx", ""xxxxx)
tweets <- get_tcs(tweet_ids, token)
}

}
