% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_token}
\alias{get_token}
\title{Get a twitter API access token}
\usage{
get_token(consumer_key = NULL, consumer_secret = NULL)
}
\arguments{
\item{consumer_key}{Character string. App consumer key.}

\item{consumer_secret}{Character string. App consumer secret.}
}
\value{
Named list containing the token.
}
\description{
Get a twitter API access token
}
