% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_tcs.R
\name{net_tcs}
\alias{net_tcs}
\title{Generate network from conversation tweets}
\usage{
net_tcs(df_tcs = NULL, type = "activity")
}
\arguments{
\item{df_tcs}{Dataframe. Conversation tweets collected by get_tcs function.}

\item{type}{Character string. Type of network to generate. Default is activity.}
}
\value{
Named list of dataframes for network nodes and edges.
}
\description{
Generate network from conversation tweets
}
\examples{
\dontrun{
# generate twitter conversation network
network <- net_tcs(tweets, "actor")

# network nodes and edges
network$nodes
network$edges
}

}
