% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcn_network.R
\name{tcn_network}
\alias{tcn_network}
\title{Generate network from conversation tweets}
\usage{
tcn_network(threads_df = NULL, type = "actor")
}
\arguments{
\item{threads_df}{Dataframe. Threaded conversation tweets collected by get_threads function.}

\item{type}{Character string. Type of network to generate. Default is "actor".}
}
\value{
Named list of dataframes for network nodes and edges.
}
\description{
Generate network from conversation tweets
}
\examples{
\dontrun{
# generate twitter conversation network
network <- tcn_network(tweets, "activity")

# network nodes and edges
network$nodes
network$edges
}

}
