% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonTCN-package.R
\docType{package}
\name{vosonTCN-package}
\alias{vosonTCN-package}
\alias{_PACKAGE}
\alias{vosonTCN}
\title{vosonTCN: Twitter Conversation Networks and Analysis}
\description{
Collects tweets and metadata for threaded conversations and generates networks.
}
\author{
\strong{Maintainer}: Bryan Gertzel \email{bryan.gertzel@anu.edu.au}

Other contributors:
\itemize{
  \item Robert Ackland \email{robert.ackland@anu.edu.au} (\href{https://orcid.org/0000-0002-0008-1766}{ORCID}) [contributor]
  \item Francisca Borquez \email{francisca.borquez@anu.edu.au} [contributor]
}

}
