% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_convo.R
\name{net_convo}
\alias{net_convo}
\title{Generate network from conversation tweets}
\usage{
net_convo(df_convo = NULL)
}
\arguments{
\item{df_convo}{Dataframe. Conversation tweets collected by get_convo function.}
}
\value{
Named list of dataframes for network nodes and edges.
}
\description{
Generate network from conversation tweets
}
\examples{
\dontrun{
# generate twitter conversation network
network <- net_convo(tweets)

# network nodes and edges
network$nodes
network$edges
}

}
