% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcn_threads.R
\name{tcn_threads}
\alias{tcn_threads}
\title{Get threaded conversation tweets}
\usage{
tcn_threads(
  tweet_ids = NULL,
  token = NULL,
  end_point = "recent",
  start_time = NULL,
  end_time = NULL,
  skip_list = NULL
)
}
\arguments{
\item{tweet_ids}{List. Tweet ids of any tweet that are part of the threaded conversations of interest. Also accepts
a list of tweet URLs or a mixed list.}

\item{token}{List. Twitter API tokens.}

\item{end_point}{Character string. Twitter API v2 search end-point. Can be either "recent" for the last 7 days or
"all" if users app has access to historical tweets. Default is "recent".}

\item{start_time}{Character string. Earliest tweet timestamp to return (UTC in ISO 8601 format). If NULL API will default to 30 days before end_time. Default is NULL.}

\item{end_time}{Character string. Latest tweet timestamp to return (UTC in ISO 8601 format). If NULL API will default to now - 30 seconds. Default is NULL.}

\item{skip_list}{Character vector. List of tweet conversation IDs to skip searching if found. This list is
automatically appended with conversation_id's when collecting multiple conversation threads to prevent search
duplication.}
}
\value{
A named list. Dataframes of tweets, users and errors.
}
\description{
Collects tweets that share the same Twitter conversation ID as supplied tweets.
}
\examples{
\dontrun{
# get twitter conversation threads by tweet ids or urls
tweet_ids <- c("xxxxxxxx",
               "https://twitter.com/xxxxxxxx/status/xxxxxxxx")
tweets <- tcn_threads(tweet_ids, token, end_point = "recent")

# get twitter conversation threads by tweet ids or urls using historical end-point
# starting from May 01, 2021.
tweet_ids <- c("xxxxxxxx",
               "https://twitter.com/xxxxxxxx/status/xxxxxxxx")
tweets <- tcn_threads(tweet_ids,
                      token = token,
                      end_point = "all",
                      start_time = "2021-05-01T00:00:00Z")
}

}
